requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
asyncpg==0.29.0  # Async PostgreSQL driver
aiosqlite==0.19.0  # SQLite async driver for testing

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2  # Pin to compatible version with passlib 1.7.4
python-multipart==0.0.6  # For file uploads

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For async testing
pytest-cov==4.1.0  # Coverage reporting

# Rate limiting (optional but included for production readiness)
slowapi==0.1.9

# Logging
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2

